# ARROW-5605: [C++] Verify Flatbuffer messages in more places to prevent crashes due to bad inputs