--- a/cpp/src/arrow/ipc/metadata-internal.cc
+++ b/cpp/src/arrow/ipc/metadata-internal.cc
@@ -1112,11 +1112,8 @@ Status GetSchema(const void* opaque_schema, DictionaryMemo* dictionary_memo,
 Status GetTensorMetadata(const Buffer& metadata, std::shared_ptr<DataType>* type,
                          std::vector<int64_t>* shape, std::vector<int64_t>* strides,
                          std::vector<std::string>* dim_names) {
-  flatbuffers::Verifier verifier(metadata.data(), metadata.size(), 128);
-  if (!flatbuf::VerifyMessageBuffer(verifier)) {
-    return Status::IOError("Verification of flatbuffer-encoded Message failed.");
-  }
-  auto message = flatbuf::GetMessage(metadata.data());
+  const flatbuf::Message* message;
+  RETURN_NOT_OK(internal::VerifyMessage(metadata.data(), metadata.size(), &message));
   auto tensor = message->header_as_Tensor();
   if (tensor == nullptr) {
     return Status::IOError("Header-type of flatbuffer-encoded Message is not Tensor.");
@@ -1150,11 +1147,8 @@ Status GetSparseTensorMetadata(const Buffer& metadata, std::shared_ptr<DataType>
                                std::vector<std::string>* dim_names,
                                int64_t* non_zero_length,
                                SparseTensorFormat::type* sparse_tensor_format_id) {
-  flatbuffers::Verifier verifier(metadata.data(), metadata.size(), 128);
-  if (!flatbuf::VerifyMessageBuffer(verifier)) {
-    return Status::IOError("Verification of flatbuffer-encoded Message failed.");
-  }
-  auto message = flatbuf::GetMessage(metadata.data());
+  const flatbuf::Message* message;
+  RETURN_NOT_OK(internal::VerifyMessage(metadata.data(), metadata.size(), &message));
   auto sparse_tensor = message->header_as_SparseTensor();
   if (sparse_tensor == nullptr) {
     return Status::IOError(