@wesm 2019-06-17
https://github.com/apache/arrow/blob/master/docs/source/developers/cpp.rst#fuzz-testing-with-libfuzzer