@crepererum 2019-06-18
Is Friday sufficient? If not, I can try to get it through tomorrow, but that's nothing I can promise.