@wesm 2019-06-21
I'm taking care of the refactoring. Done shortly